% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{bin.variables}
\alias{bin.variables}
\title{Bin them vars}
\usage{
bin.variables(given, breaks, labels, data, bins, ghost.reference,
  prediction)
}
\arguments{
\item{given}{stuff}

\item{breaks}{The breaks to be used for the bins}

\item{labels}{A list of the same length as bins}

\item{data}{the dataset}

\item{bins}{The number of bins used when putting quantitative variables into categories. A list can be used if different amounts are wanted for different variables}

\item{ghost.reference}{What should the reference group be (from which to draw the ghost line)? The user can specify up to two values as a vector. See examples.}

\item{prediction}{The predicted line}
}
\description{
Bin them vars
}
\details{
Bin them vars
}
\author{
Dustin Fife
}
