% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.plot.R
\name{bivariate.plot}
\alias{bivariate.plot}
\title{Plot bivariate plot of two variables}
\usage{
bivariate.plot(x, y, x.numeric = NULL, y.numeric = NULL, d = NULL, ...)
}
\arguments{
\item{x}{A string indicating the name of the IV to plot. Alternatively, an object can be used.}

\item{y}{A string indicating the name of the DV to plot. Alternatively, an object can be used.}

\item{x.numeric}{Logical. Is the predictor variable of interest numeric?}

\item{y.numeric}{Logical. Is the outcome variable of interest numeric?}

\item{...}{Other arguments passed to ggplot functions}
}
\value{
A plot
}
\description{
A function that automatically plots either a scatterplot, mean plot, or a barchart
}
\details{
Ideally, the user will provide the data type for this function (by saying numeric = T, or numeric = F)
Omitting that, the function will decide for the user. If both variables are numeric (or if there are more than five unique values of the variable)
it will generate a scatterplot. If one variable is a factor OR it has five or less unique values, it will output a mean plot. If all
variables are categorical, it will output a barchart.
}
\examples{
x = rnorm(100)
y = rnorm(100)
bivariate.plot(x,y)
}
\author{
Dustin Fife
}
\seealso{
\code{\link{uni.plot}}
}

