% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.plot.R
\name{chisq.plot}
\alias{chisq.plot}
\title{Plot Expected and Observed Frequencies (Chi Square)}
\usage{
chisq.plot(list = NULL, v1 = NULL, v2 = NULL, standardize = F)
}
\arguments{
\item{list}{A table containing the variables. If not included, user can include raw data as v1 or v2}

\item{v1}{The first variable (must be a factor)}

\item{v2}{The second variable (also must be a factor)}

\item{standardize}{Logical. Should the results be reported in proportions? (Rather than counts)}
}
\description{
This function will plot the expected vs. observed frequencies from a Chi Square Test and plot them in Barplots.
}
\examples{
##' 
# From Agresti(2007) p.39
M = as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) = list(gender = c("F", "M"),
                    party = c("Democrat","Independent", "Republican"))
chisq.plot(list=M, standardize=F)              
}
\author{
Dustin Fife
}
\references{
Agresti, A. (2007) An Introduction to Categorical Data Analysis, 2nd ed., New York: John Wiley & Sons. Page 38.
}
\seealso{
\code{\link{barchart}}
}

