% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.fits.R
\name{compare.fits}
\alias{compare.fits}
\title{Compare the fits of two models}
\usage{
compare.fits(formula, data, model1, model2 = NULL, return.preds = F,
  ...)
}
\arguments{
\item{formula}{A formula that can be used in flexplot. The variables inside must not include variables outside the fitted models.}

\item{data}{The dataset containing the variables in formula}

\item{model1}{The fitted model object (e.g., lm) containing the variables specified in the formula}

\item{model2}{The second fitted model object (e.g., lm) containing the variables specified in the formula}
}
\description{
Compare the fits of two models
}
\details{
Compare the fits of two models
}
\examples{
#not yet
}
\author{
Dustin Fife
}
