% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarCoord.R
\name{ellipse}
\alias{ellipse}
\title{Plots an Ellipse}
\usage{
ellipse(x0, y0, axisX, axisY, color = "lightgray")
}
\arguments{
\item{x0}{the x coordinate of the center of the circle}

\item{y0}{the y coordinate of the center of the circle}

\item{axisX}{the radius along the x axis}

\item{axisY}{the radius along the y axis}

\item{color}{what color should the ellipse be drawn in}
}
\description{
##' Plots an ellipse
}
\author{
Dustin Fife
}

