% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{estimates.fifer.anova}
\alias{estimates.fifer.anova}
\alias{fifer.anova.estimates}
\title{Report fifer.anova Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{fifer.anova}(object)
}
\arguments{
\item{object}{a fifer.anova object}
}
\description{
Report fifer.anova object Estimates
}
