% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.lm}
\alias{estimates}
\alias{estimates.lm}
\title{Report regression object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{lm}(object)
}
\arguments{
\item{object}{a lm object}
}
\description{
Report regression object Estimates
}

