% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{estimates.ttest}
\alias{estimates.ttest}
\alias{estimates}
\title{Report ttest object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{ttest}(object)
}
\arguments{
\item{object}{a ttest object}
}
\description{
Report ttest object Estimates
}
