% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_correlation.R
\name{exact_correlation}
\alias{exact_correlation}
\title{Simulate data with exact correlations}
\usage{
exact_correlation(n, rho)
}
\arguments{
\item{n}{The sample size}

\item{rho}{The desired correlation coefficient}
}
\value{
A matrix containing correlations between the two variables
}
\description{
Most algorithms for simulating random normal variables of a particular correlation
(e.g., \eqn{\rho=0.5}) only do so \emph{on average}. This function will generate random normal
data with exactly the correlation specified.
}
\examples{
cor(exact_correlation(100, .5))
cor(exact_correlation(5, 0))
}
\references{
See \url{https://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variables}
}
