% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.plot.R
\name{factorial.plot}
\alias{factorial.plot}
\title{Plot a Factorial ANOVA}
\usage{
factorial.plot(y, x1, x2, yname = NULL, x1name = NULL, x2name = NULL)
}
\arguments{
\item{y}{The outcome of interest}

\item{x1}{The first factor. This will be plotted on the x-axis}

\item{x2}{The second factor. This will be plotted in the columns of a plot grid}
}
\value{
A plot
}
\description{
Plot a Factorial ANOVA
}
\details{
This function plots the distribution of Y scores conditional on two factors. The first factor will occupy the x-axis, while the second will be displayed using different plots.
}
\author{
Dustin Fife
}

