% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.tests.R
\docType{data}
\name{fakeMedicalData}
\alias{fakeMedicalData}
\title{Fictitious Medical Dataset}
\description{
Fictitious Medical Dataset
}
\author{
Dustin Fife \email{blahblah@roxygen.org}
}
\keyword{data}

