% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{fifer.anova}
\alias{fifer.anova}
\title{Perform a one-way anova}
\usage{
fifer.anova(y, group)
}
\arguments{
\item{y}{A vector containing the Dependent variable scores}

\item{x}{A vector containing the group categories}
}
\description{
This function performs a one-way anova, but does so in a way that adheres to the 7 steps of data analysis (it reports residuals/effect sizes/parameter estimates before showing significance).
}
\details{
The validity of an anova relies on basic statistical assumptions (normality, homoskedasiticity, and independence). Furthermore, statistical significance can easily be
conflated with practical significance. This function is simply a wrapper for r's native \code{\link{anova}} function, but outputs the results in such a way that encourages
the user to focus on estimation as model appropriateness.
}
\examples{
x = rep(1:3, times=30)
y = 10 + 5*x + rnorm(90, 0, 8)
a = fifer.anova(y,group=x)
plot(a)
estimates(a)
report(a)
}
\seealso{
\code{\link{anova}}
}
\author{
Dustin Fife
}
