% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexplot.R
\name{flexplot}
\alias{flexplot}
\title{Create a "flexible plot" or flexplot}
\usage{
flexplot(formula, data, color = NULL, symbol = NULL, linetype = NULL,
  bins = 4, labels = NULL, breaks = NULL, method = "loess", se = T,
  spread = c("quartiles", "stdev", "sterr"), jitter = FALSE, raw.data = T,
  sample = Inf)
}
\arguments{
\item{formula}{A formula of the form  y~x + a | b + z}

\item{data}{The dataset}

\item{color}{Color values to be used for the categorical variables}

\item{symbol}{The symbols to be used for the categorical variables}

\item{linetype}{The linetype to be used for the categorical variables}

\item{bins}{The number of bins used when putting quantitative variables into categories. A list can be used if different amounts are wanted for different variables}

\item{labels}{A list of the same length as bins}

\item{breaks}{The breaks to be used for the bins}

\item{method}{The method to be used to draw the lines. Defaults to loess}

\item{se}{Should standard errors be drawn?}

\item{spread}{How should standard errors be drawn? Defaults to quartiles}

\item{jitter}{Should values be jittered?}

\item{raw.data}{Should raw data be plotted?}

\item{sample}{Should a sample of the data be plotted? Defaults to Inf (for all variables).}
}
\description{
Create a flexible plot
}
\details{
Formula takes the form of y~x + a | b + z. Everything on the right of | will occur in panels (in this case, b will be in the columns and z will be in the rows). Any numeric
variables between ~ and | will fall on the x axis and categorical predictors will be color-coded/symbol-coded/line-coded (if user specifies). If more than one numeric variable is specified
after ~, the second (and third) will be binned. No more than four variables are allow (otherwise, cognitive load is too high). If the user wishes to include another variable,
 I recommend they create separate plots, one for each level of the new variable (or a binned level of a numeric variable).
}
\examples{

}
\author{
Dustin Fife
}

