% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.data}
\alias{generate.data}
\title{Simulate multivariate normal data}
\usage{
generate.data(
  cor.matrix,
  sds = c(1, 1),
  means = c(0, 0),
  names = c("y", "x"),
  n = 100
)
}
\arguments{
\item{cor.matrix}{Either a scalar (indicating the correlation between two variables) or a symmetrical matrix for a variance/covariance matrix}

\item{sds}{The standard deviation of the variables}

\item{means}{The means of the variables}

\item{names}{The names of the variables}

\item{n}{The desired sample size}
}
\value{
A data frame containing the simulated data
}
\description{
Simulate multivariate normal data
}
\details{
This function provides an easy way to simulate random multivariate normal data.
}
\examples{
generate.data(cor.matrix=.2, sds=c(10, 100), means=c(30, 1000), c("A", "B"), 200)
}
\author{
Dustin Fife
}
