% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_polynomial.R
\name{generate_polynomial}
\alias{generate_polynomial}
\title{Generate polynomial random data}
\usage{
generate_polynomial(
  cor = 0.3,
  poly = -0.4,
  n = 200,
  means = c(0, 0),
  sds = c(1, 1),
  names = c("x", "y"),
  plot = TRUE,
  ...
)
}
\arguments{
\item{cor}{correlation coefficient}

\item{poly}{strength of polynomial term}

\item{n}{sample size}

\item{means}{means of x and y, respectively}

\item{sds}{sds of x and y, respectively}

\item{names}{names of x/y, respectively}

\item{plot}{should a plot be returned?}

\item{...}{other arguments passed to flexplot}
}
\value{
either a plot or generated data
}
\description{
Generate polynomial random data
}
