% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve.fit.R
\name{improve.fit}
\alias{improve.fit}
\title{Compare the improvement in fit from a baseline model}
\usage{
improve.fit(full, reduced)
}
\arguments{
\item{full}{the full mixed model}

\item{reduced}{the reduced mixed model (typically a random effect ANOVA model)}
}
\value{
A vector containing the proportion change in fit for each random effects parameter
}
\description{
Compare the improvement in fit from a baseline model
}
\details{
Compare the improvement in fit from a baseline model, using mixed models
}
\examples{
data(math)
require(lme4)
baseline.model = lmer(MathAch~1 + (1|School), data=d)
full.model = lmer(MathAch~SES + (SES|School), data=d)
improve.fit(full.model, baseline.model)
}
\author{
Dustin Fife
}
