% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.me.R
\name{impute.me}
\alias{impute.me}
\title{Multiple Imputation on a Model}
\usage{
impute.me(model, data, predictors = NULL, keep = T, imputations = 20,
  silent = F)
}
\arguments{
\item{model}{An R-friendly model. Currently, it only allows lm objects, but will eventually allow other objects (e.g., glm).}

\item{data}{The dataset used for analysis. This dataset should contain predictors used to impute the missing values}

\item{predictors}{A list of predictors (as a character vector) that identify which variables to keep (or drop; see below argument).}

\item{keep}{Logical. Should the list of predictors be kept or dropped? Defaults to keep.}

\item{imputations}{The number of imputations to be performed. Defaults to 20.}
}
\description{
Multiple Imputation on a Model
}
\details{
This is a wrapper function for both the mice function in the mice package, as well as for basic models in R (e.g., lm). As input,
it takes the model the user wishes to estimate using advanced missing data strategies, as well as a list of variables they wish to use
to impute the missing values. The function takes the raw data and performs MI using mice, then re-analyzes the dataset and outputs the
multiply imputed parameter estimates.
}
\examples{
data(exercise_data)
d = exercise_data

		##### create missing data in motivation
missing.ld = which(d$motivation<quantile(d$motivation, .25))
notmissing = which(!(1:nrow(d) \%in\% missing.ld))
d$weight.change.missing = d$weight.change
d$weight.change.missing[missing.ld] = NA

		#### create model with missing data
mod = lm(weight.change.missing~motivation, data=d, model=T)
predictors = c("muscle.gain.missing", "weight.change")
impute.me(mod, data=d, predictors=predictors, keep=F, imputations=5)
}
\author{
Dustin Fife
}
