% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.me.R
\name{impute.model.comparison}
\alias{impute.model.comparison}
\title{Compute Bayes Factor for a Imputed Model}
\usage{
impute.model.comparison(
  model1,
  model2,
  data,
  predictors = NULL,
  keep = T,
  imputations = 20,
  silent = F,
  invert = F
)
}
\arguments{
\item{model1}{The full model}

\item{model2}{The reduced model}

\item{data}{The dataset used for analysis. This dataset should contain predictors used to impute the missing values}

\item{predictors}{A list of predictors (as a character vector) that identify which variables to keep (or drop; see below argument).}

\item{keep}{Logical. Should the list of predictors be kept or dropped? Defaults to keep.}

\item{imputations}{The number of imputations to be performed. Defaults to 20.}
}
\value{
.mod Should the model be returned?
}
\description{
Compute Bayes Factor for a Imputed Model
}
\details{
blah blah blah
}
\author{
Dustin Fife
}
