% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersperse.R
\name{intersperse}
\alias{intersperse}
\title{Intersperse elements of Two+ Vectors}
\usage{
intersperse(...)
}
\arguments{
\item{...}{the vectors the user wishes to intersperse}
}
\value{
a vector of length i X j, containing the interspersed vectors as strings
}
\description{
Intersperse elements of two+ vectors
Given two vectors (one of length i, the other of length j), \code{intersperse} will combine the elements of each vector into strings of length
i X j, where each element is the concatenation of the elements of the two vectors. See examples.
}
\examples{
intersperse(LETTERS[1:3], 1:3)
}
\author{
Dustin Fife
}

