% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{jit.func}
\alias{jit.func}
\title{How to display data}
\usage{
jit.func(jitter, data, ss, raw.data, alpha)
}
\arguments{
\item{jitter}{Should data be jittered?}

\item{data}{the dataset}

\item{ss}{The size of the sample}

\item{raw.data}{should raw data be displayed}

\item{alpha}{the transparency level}
}
\description{
How to display data
}
\details{
How to display data
}
\author{
Dustin Fife
}
