% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.symmetric.R
\name{make.symmetric}
\alias{make.symmetric}
\title{Force a matrix to be symmetric}
\usage{
make.symmetric(a, lower.tri = TRUE)
}
\arguments{
\item{a}{A matrix you wish to force to be symmetrical}

\item{lower.tri}{Should the upper triangle be replaced with the lower triangle?}
}
\value{
a symmetric matrix
}
\description{
Force a matrix to be symmetric
}
\examples{
a = matrix(rnorm(16), ncol=4)
make.symmetric(a, lower.tri=FALSE)
}
\author{
Dustin Fife
}

