\name{missing.vals}
\alias{missing.vals}
\title{Summary of Missing Data}
\usage{
  missing.vals(dataset)
}
\arguments{
  \item{dataset}{a matrix (persons in rows, variables in
  columns).}
}
\value{
  a matrix with information about what is missing.
}
\description{
  Provides a summary of missing data.
}
\author{
  Dustin Fife
}

