% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{plot.fifer.anova}
\alias{plot.fifer.anova}
\alias{plot}
\title{Plot fifer.anova Summary}
\usage{
\method{plot}{fifer.anova}(x, ...)
}
\arguments{
\item{x}{a fifer.anova object}

\item{...}{other parameters passed to plot}

\item{y}{igorned}
}
\description{
Plot fifer.anova Summary
}
