% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{plot.regression}
\alias{plot.regression}
\alias{plot}
\title{Plot regression Summary}
\usage{
\method{plot}{regression}(x, ...)
}
\arguments{
\item{x}{a regression object}

\item{...}{other parameters passed to plot}

\item{y}{igorned}
}
\description{
Plot regression Summary
}
