% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.rfPred}
\alias{plot.rfPred}
\title{Plot rfPred Summary}
\usage{
\method{plot}{rfPred}(x, y, ...)
}
\arguments{
\item{x}{an rfPred object}

\item{y}{igorned}

\item{...}{other parameters passed to plo}
}
\description{
Plot rfPred Summary
}

