\name{plot.rfThresh}
\alias{plot.rfThresh}
\title{Plot rf.thesh Summary}
\usage{
\method{plot}{rfThresh}(x, y, ...)

\method{plot}{rfThresh}(x, y, ...)
}
\arguments{
  \item{x}{an rfThresh object}

  \item{y}{igorned}

  \item{...}{other parameters passed to plo}

  \item{x}{an rfThresh object}

  \item{y}{igorned}

  \item{...}{other parameters passed to plo}
}
\description{
Plot rf.thesh Summary

Plot rf.thesh Summary
}

