% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{plot.ttest}
\alias{plot.ttest}
\alias{plot}
\title{Plot ttest Summary}
\usage{
\method{plot}{ttest}(x, ...)
}
\arguments{
\item{x}{a ttest object}

\item{...}{other parameters passed to plot}

\item{y}{igorned}
}
\description{
Plot ttest Summary
}
