% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted.line.R
\name{predicted.line}
\alias{predicted.line}
\title{Create and plot a fitted line for a generalized linear model}
\usage{
predicted.line(
  model,
  outcome,
  IV,
  covs = NA,
  data,
  cov.func = c("mean", "median", "min", "max"),
  length = 100,
  add = F,
  ret.pred = F,
  ...
)
}
\arguments{
\item{model}{A model (e.g., lmer model, or any type of model that supports the "predict" command)}

\item{outcome}{A string that names the outcome variable}

\item{IV}{A string that names the predictor variable}

\item{data}{The original dataset used to fit the model}

\item{length}{The number of predicted points used for plotting the line. Defaults to 100.}

\item{ret.pred}{Should the fits be returned as a matrix?}

\item{...}{Other arguments passed to plot, lines, or predict}

\item{covariates}{A function used on all the covariates. Can be mean, median, min, or max}

\item{Add.}{Should the fitted line be added to an existing plot?}
}
\description{
Create and plot a fitted line for a generalized linear model
}
\examples{
## do this later
}
\author{
Dustin Fife
}
