% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{print.fifer.anova}
\alias{print.fifer.anova}
\title{Print fifer.anova Summary}
\usage{
\method{print}{fifer.anova}(x, ...)
}
\arguments{
\item{x}{a fifer.anova object}

\item{...}{ignored}
}
\description{
Print fifer.anova Summary
}
