% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{print.regression}
\alias{print.regression}
\alias{regression.ttest}
\title{Print regression Summary}
\usage{
\method{print}{regression}(x, ...)
}
\arguments{
\item{x}{a regression object}

\item{...}{ignored}
}
\description{
Print regression Summary
}

