% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.interp.R
\name{print.rfInterp}
\alias{print.rfInterp}
\title{Print rfInterp Summary}
\usage{
\method{print}{rfInterp}(x, ...)
}
\arguments{
\item{x}{an rfInterp object}

\item{...}{ignored}
}
\description{
Print rfInterp Summary
}

