\name{print.rfThresh}
\alias{print.rfThresh}
\title{Print rfThesh Summary}
\usage{
\method{print}{rfThresh}(x, ...)

\method{print}{rfThresh}(x, ...)
}
\arguments{
  \item{x}{an rfThresh object}

  \item{...}{ignored}

  \item{x}{an rfThresh object}

  \item{...}{ignored}
}
\description{
Print rfThesh Summary

Print rfThesh Summary
}

