\name{printx}
\alias{printx}
\title{Change Defaults of print.xtable}
\usage{
  printx(x, file = "", ...)
}
\arguments{
  \item{file}{the location where the file is stored}

  \item{...}{other arguments passed to print.xtable}
}
\description{
  Change Defaults for print.xtable
}
\author{
  Dustin Fife
}

