\name{random.correlation}
\alias{random.correlation}
\title{Generate a random correlation matrix}
\usage{
  random.correlation(n)
}
\arguments{
  \item{n}{the number of rows/dimensions of the correlation
  matrix}
}
\value{
  a correlation matrix, of size \code{nxn}
}
\description{
  Generate a random correlation matrix
}
\author{
  Dustin Fife
}

