% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{raw.alph.func}
\alias{raw.alph.func}
\title{Find transparency of data}
\usage{
raw.alph.func(raw.data, alpha = 1)
}
\arguments{
\item{raw.data}{Should raw data be displayed?}

\item{alpha}{the transparency level}
}
\description{
Find transparency of data
}
\details{
Find transparency of data
}
\author{
Dustin Fife
}
