% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{Perform a regression/correlation analysis}
\usage{
regression(y, x)
}
\arguments{
\item{y}{The outcome variable of interest}

\item{x}{The predictor variable of interest}
}
\description{
Perform a regression/correlation analysis in such a way that it focuses on statistics of interest (graphics first, then estimates, then p-values)
}
\details{
This function first shows a plot of the relationship, along with residual plots. When requested, it also outputs estimates and p-values
}
\examples{
x = rnorm(100, 0, 1)
y = .8*x + rnorm(100, 0, sqrt(1-.8^2))
reg.mod = regression(x,y)
estimates(reg.mod)
plot(reg.mod)
report(reg.mod)
}
\seealso{
\code{\link{MASS}}
}
\author{
Dustin Fife
}
