% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{related.plot}
\alias{related.plot}
\title{Function to create a related t test plot}
\usage{
related.plot(predictors, outcome, data, raw.data, ss, spread = "quartiles")
}
\arguments{
\item{predictors}{the character of the predictor variable}

\item{outcome}{the character of the outcome variable}

\item{data}{the dataset}

\item{raw.data}{should raw data be displayed}

\item{spread}{type of spread to be displayed}

\item{n}{the number of datapoints to be displayed}
}
\description{
Function to create a related t test plot
}
\details{
Function to create a related t test plot
}
\author{
Dustin Fife
}
