% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace}
\alias{replace}
\title{Replace values in a vector}
\usage{
replace(x, old.value, new.value, factor = F)
}
\arguments{
\item{x}{The vector that has values that need replacing}

\item{old.value}{The value that needs replacing}

\item{new.value}{The value that will replace old.value}

\item{factor}{Should the vector be "refactored" (so old values no longer show up as a factor level)?}
}
\value{
A new vector where old.values are replaced by new.values
}
\description{
Replace the values in a vector with a new value
}
\details{
This function searches the vector provided for old.values, then replaces each occurance with new.values
}
\examples{
d = data.frame(category = sample(c("Male", "Female", "Men"), prob=c(.45, .45, .05), size=100, replace=T))
table(d$category)
d$category = replace(d$category, "Men", "Male")
table(d$category)
}
\author{
Dustin Fife
}
