% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{report.regression}
\alias{report.regression}
\alias{report}
\alias{report.default}
\title{Output APA style statistical significance from a regression object}
\usage{
\method{report}{regression}(object)
}
\arguments{
\item{object}{a regression object}
}
\description{
Output APA style statistical significance from a regression object
}
