% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{report.ttest}
\alias{report}
\alias{report.default}
\alias{report.ttest}
\title{Output APA style statistical significance from a ttest object}
\usage{
\method{report}{ttest}(object)
}
\arguments{
\item{object}{a ttest object}
}
\description{
Output APA style statistical significance from a ttest object
}

