% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualize.lowess.R
\name{residualize.lowess}
\alias{residualize.lowess}
\title{Residualize a lowess function}
\usage{
residualize.lowess(x, y)
}
\arguments{
\item{x}{The predictor of interest}

\item{y}{The outcome of interest}
}
\value{
A list of residuals
}
\description{
This function will create an approximate fitted function for a lowess line, then residualize the Y based on the fit of the model
}
\examples{
data(airquality)
head(airquality)
attach(airquality)
plot(Wind, Temp)
lines(lowess(Wind, Temp), col="red")
residuals = residualize.lowess(Wind, Temp)
plot(Wind, residuals)
}
\author{
Dustin Fife
}
\references{
This function was developed by Glen_b on \href{https://stats.stackexchange.com/questions/126699/residuals-from-lowess-curve}{Stack Overflow}
}

