% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rfSensitivity}
\alias{rfSensitivity}
\title{Output accuracy, sensitivity, specificity, NPV, and PPV.}
\usage{
rfSensitivity(object)
}
\arguments{
\item{object}{An rfPred object}
}
\value{
A list containing the accuracy, sensitivity, etc.
}
\description{
Output accuracy, sensitivity, specificity, NPV, and PPV.
}

