% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot for 1080p}
\usage{
save_plot(plot, file_location = "", file_name = "plot", text.size = 18)
}
\arguments{
\item{plot}{a ggplot2 object}

\item{file_location}{the location where the file will be saved}

\item{file_name}{the name of the file}

\item{text.size}{the size of the font. Defaults to 18}
}
\value{
a plot
}
\description{
Save a plot for 1080p
}
