% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.it.R
\name{scaleIt}
\alias{scaleIt}
\title{Scale a variable}
\usage{
scaleIt(x, mean = NULL, sd = NULL, min = NULL, max = NULL)
}
\arguments{
\item{x}{the variable to be scaled}

\item{mean}{the mean you wish the distribution to have}

\item{sd}{the sd you wish the distribution to have}

\item{min}{the min you wish the distribution to have}

\item{max}{the max you wish the distribution to have}
}
\value{
the scaled variable
}
\description{
Scale a variable to have a particular mean/sd (or min/max)
}
\author{
Dustin Fife
}

