% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{spread_type}
\alias{spread_type}
\title{Type of spread to use}
\usage{
spread_type(spread, predictors, outcome, data)
}
\arguments{
\item{spread}{Should be stdev, sterr, or quartiles}

\item{predictors}{the character of the predictor variable}

\item{outcome}{the character of the outcome variable}

\item{data}{the dataset}
}
\description{
Type of spread to use
}
\details{
Type of spread to use
}
\author{
Dustin Fife
}
