% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized.beta.R
\name{standardized.beta}
\alias{standardized.beta}
\title{Compute standardized betas on a linear model object}
\usage{
standardized.beta(object)
}
\arguments{
\item{object}{a lm object}
}
\value{
the standardized betas
}
\description{
Compute standardized betas on a linear model object
}
\details{
Compute standardized betas on a linear model object
}
\author{
Dustin Fife
}
