% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{subset.sample}
\alias{subset.sample}
\title{Sample the data}
\usage{
\method{subset}{sample}(ss, data)
}
\arguments{
\item{ss}{sample size of raw data to be displayed. Defaults to Inf (for all data)}

\item{data}{the dataset}
}
\description{
Sample the data
}
\details{
Sample the data
}
\author{
Dustin Fife
}
