% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.pred.R
\name{summary.rfPred}
\alias{summary.rfPred}
\title{Print a Summary Table of rfPred}
\usage{
\method{summary}{rfPred}(object, ...)
}
\arguments{
\item{object}{an rfPred object}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
summary.rfPred is best for those non-Latex users to produce a table
that shows each stage of the variable selection algorithm.
}
\details{
Print a Summary Table of rfPred
}

