% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{Perform a two-independent sample t-test}
\usage{
ttest(y, x)
}
\arguments{
\item{y}{Either a vector containing the Dependent variable scores, or a vector containing the scores of group 1}

\item{x}{Either a vector containing the group categories, or a vector containing the scores of group 2}
}
\value{
Two objects: cohen's d and a table of estimates (means and difference between groups)
}
\description{
This function performs a t-test, but does so in a way that adheres to the 7 steps of data analysis (it reports residuals/effect sizes/parameter estimates before showing significance).
}
\details{
The validity of a t-test relies on basic statistical assumptions (normality, homoskedasiticity, and independence). Furthermore, statistical significance can easily be
conflated with practical significance. This function is simply a wrapper for r's native \code{\link{t.test}} function, but outputs the results in such a way that encourages
the user to focus on estimation as model appropriateness.
}
\examples{
# where y and x are scores for group 1 and group 2 (respectively)
y = rnorm(30, 10, 5)
x = rnorm(30, 12, 5)
ttest(y,x)

# where y is all scores and x is the group labels
y = rnorm(300, 10, 5)
x = sample(c(1:2), size=length(y), replace=T)
ttest(y,x)
}
\author{
Dustin Fife
}
\seealso{
\code{\link{t.test}}
}

