% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.export.R
\name{video.export}
\alias{video.export}
\title{Export to video format a ggplot image}
\usage{
video.export(a, name, folder = "images", shape.plot = "horizontal")
}
\arguments{
\item{a}{a ggplot object to be exported}

\item{name}{the name of the file}

\item{folder}{the directory in images where the file is to be placed. This defaults to "images" in the root folder.}

\item{shape.plot}{the shape of the plot for scaling. Can be "vertical", "horizontal", or "square"}
}
\description{
This will take a ggplot2 object and save it with the appropriate settings for a 1080p video. 
It also attempts to maintain correct aspect ratios in the process (ideally).
}
\details{
Export to video format a ggplot image
}
\examples{
d = data.frame(Y=rnorm(10000, 50, 10))
a = flexplot::flexplot(Y~1, data=d)
video.export(a, "histogram", "distributions")
}
\references{
There's some diffulty with setting a fixed size: 
\url{https://stackoverflow.com/questions/54892086/setting-canvas-size-in-ggsave-to-a-fixed-width-while-maintaining-aspect-ratio-of}
}
