% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.lm}
\alias{visualize.lm}
\alias{visualize}
\alias{visualize.default}
\title{Visualize a fitted model}
\usage{
\method{visualize}{lm}(object, plot = c("all", "residuals", "bivariate"),
  linetype = "loess", formula = NULL, ...)
}
\arguments{
\item{object}{a lm object}

\item{plot}{what should be plotted? Residuals? Bivariate plot? All of them?}

\item{linetype}{What time of smoothing line should be drawn? Defaults to loess.}

\item{formula}{A flexplot-style formula}
}
\description{
Visualize a fitted model
}
