% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{x.summary}
\alias{x.summary}
\title{Type of x summary to use (line or mean)}
\usage{
x.summary(predictors, outcome, spread, data)
}
\arguments{
\item{predictors}{the character of the predictor variable}

\item{outcome}{the character of the outcome variable}

\item{spread}{type of spread to be displayed}

\item{data}{the dataset}
}
\description{
Type of x summary to use (line or mean)
}
\details{
Type of x summary to use (line or mean)
}
\author{
Dustin Fife
}
