% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarCoord.R
\name{compute.theta}
\alias{calcT}
\alias{compute.theta}
\title{Calculate inverse tangent.}
\usage{
compute.theta(x)
}
\arguments{
\item{x}{a vector of size two that give the cartesian coordinates}
}
\value{
returns the theta, expressed in radians.
}
\description{
Computes angle in polar coordinates, accounting for which quadrant the datapoints are in.
}
\author{
Dustin Fife
}

