% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{contents}
\alias{contents}
\title{View the contents (functions) of a package}
\usage{
contents(x = NULL)
}
\arguments{
\item{x}{a string containing the name of the package. Defaults to "fifer".}
}
\value{
a string of all the functions contained within that package.
}
\description{
View the contents (functions) of a package
}
\details{
It's easy to forget what functions are contained within a package.
\code{contents} will return all the functions within a package
}
\examples{
## see what's in fifer
contents()
## see what's in MASS
\dontrun{contents("MASS")}
}
\author{
Dustin Fife
}

