% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R, R/regression.R, R/visualize.R
\name{estimates.default}
\alias{estimates.default}
\alias{report.regression}
\alias{report}
\alias{report.ttest}
\alias{report.default}
\alias{visualize}
\alias{visualize.lm}
\alias{visualize.default}
\title{Output APA style statistical significance from an object}
\usage{
\method{estimates}{default}(object, plot = c("all", "residuals",
  "bivariate"))

\method{estimates}{default}(object, plot = c("all", "residuals",
  "bivariate"))

\method{estimates}{default}(object, plot = c("all", "residuals",
  "bivariate"))
}
\arguments{
\item{object}{a object}

\item{plot}{what should be plotted? Residuals? Bivariate plot? All of them?}

\item{object}{a object}

\item{object}{a object}

\item{formula}{A flexplot-style formula}

\item{linetype}{What time of smoothing line should be drawn? Defaults to loess.}
}
\description{
Output APA style statistical significance from an object

Output APA style statistical significance from an object

Visualize a fitted model
}
