% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{estimates.default}
\alias{estimates.default}
\alias{report}
\alias{report.default}
\alias{report.regression}
\alias{report.ttest}
\title{Output APA style statistical significance from an object}
\usage{
\method{estimates}{default}(object)
}
\arguments{
\item{object}{a object}
}
\description{
Output APA style statistical significance from an object
}

