% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{estimates}
\alias{estimates}
\alias{estimates.default}
\alias{estimates.regression}
\title{Report regression object Estimates (effect sizes and parameters)}
\usage{
estimates(object, ...)
}
\arguments{
\item{object}{a regression object}
}
\description{
Report regression object Estimates
}

