% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.null.R
\name{get.cols}
\alias{get.cols}
\title{Extract column index}
\usage{
get.cols(string, names)
}
\arguments{
\item{string}{the name of a variable (string)}

\item{names}{the names of the variables for which you wish to extract the column}
}
\value{
an integer corresponding to the column index(indices)
}
\description{
Given a vector of strings and another reference vector, \code{get.cols} will search the reference
vector for matches and return the column indices where each string was located. See examples.
}
\examples{
names = LETTERS[1:10]
get.cols(c("A", "B", "D"), LETTERS[1:10])  ### works
#get.cols(c("A", "B", "Z"), LETTERS[1:10])  ### doesn't work
}
\author{
Dustin Fife
}
\seealso{
\code{\link{make.null}}, \code{\link{r}}
}

