% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.vals.R
\name{missing.vals}
\alias{missing.Vals}
\alias{missing.vals}
\alias{missingVals}
\alias{missingvals}
\title{Summary of Missing Data}
\usage{
missing.vals(dataset)
}
\arguments{
\item{dataset}{a matrix or data frame (persons in rows, variables in columns).}
}
\value{
a matrix with information about what is missing.
}
\description{
Given a dataset, \code{missing.vals} will report on the number
of missing observations per variable
}
\author{
Dustin Fife
}

