% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.interp.R
\name{plot.rfInterp}
\alias{plot}
\alias{plot.rfInterp}
\title{Plot rfInterp Summary}
\usage{
\method{plot}{rfInterp}(x, y, ...)
}
\arguments{
\item{x}{an rfInterp object}

\item{y}{igorned}

\item{...}{other parameters passed to plot}
}
\description{
Plot rfInterp Summary
}

