% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.pred.R
\name{plot.rfPred}
\alias{plot.rfPred}
\title{Plot rfPred Summary}
\usage{
\method{plot}{rfPred}(x, y, ...)
}
\arguments{
\item{x}{an rfPred object}

\item{y}{igorned}

\item{...}{other parameters passed to plo}
}
\description{
Plot rfPred Summary
}

