% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.thresh.R
\name{plot.rfThresh}
\alias{plot.rfThresh}
\title{Plot rf.thesh Summary}
\usage{
\method{plot}{rfThresh}(x, y, ...)
}
\arguments{
\item{x}{an rfThresh object}

\item{y}{igorned}

\item{...}{other parameters passed to plo}
}

