% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.rfThresh}
\alias{plot.rfThresh}
\title{Plot rf.thesh Summary}
\usage{
\method{plot}{rfThresh}(x, y, ...)
}
\arguments{
\item{x}{an rfThresh object}

\item{y}{igorned}

\item{...}{other parameters passed to plo}
}
\description{
Plot rf.thesh Summary
}

