\name{plotSigBars}
\alias{plotSigBars}
\title{Add significance bars to a prism plot}
\usage{
  plotSigBars(formula, data, type = "tukey")
}
\arguments{
  \item{formula}{a R formula object}

  \item{data}{a dataset containing the variables in
  formula}

  \item{type}{either "tukey" or "dunn" indicating which
  multiple comparison should be used}
}
\description{
  Add significance bars to a prism plot
}
\author{
  Dustin Fife
}
\seealso{
  \code{\link{boxplot}}, \code{\link{densityPlotR}},
  \code{\link{prism.plots}}
}

