% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.thresh.R
\name{print.rfThresh}
\alias{print.rfThresh}
\title{Print rfThesh Summary}
\usage{
\method{print}{rfThresh}(x, ...)
}
\arguments{
\item{x}{an rfThresh object}

\item{...}{ignored}
}
\description{
Print a rfThresh object
}

