% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printx.R
\name{printx}
\alias{printx}
\title{Change Defaults of print.xtable}
\usage{
printx(x, file = "", ...)
}
\arguments{
\item{x}{The \code{xtable} object the user wishes to export}

\item{file}{the location where the file is stored}

\item{...}{other arguments passed to print.xtable}
}
\description{
Change Defaults for print.xtable
}
\details{
This is a wrapper for the function \code{\link{print.xtable}}, where the defaults have been specified
for apa-like tables.
}
\author{
Dustin Fife
}

