% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.null.R
\name{r}
\alias{r}
\title{Extract variable column indices}
\usage{
r(stringA, stringB, data.names, names = F)
}
\arguments{
\item{stringA}{the name of the first variable you wish to extract the column for in sequence}

\item{stringB}{the name of the last variable you wish to extract the column for in sequence}

\item{data.names}{the names of the dataset}

\item{names}{Should the names of the variables be returned? Or the column indices? Defaults to FALSE 
(meaning the column indices will be returned).}
}
\value{
a vector of numbers corresponding to the column names
}
\description{
Given a starting string and an ending string, \code{r} will extract all columns between the first 
and the last string and return either a vector of strings or a vector of integers.
See examples.
}
\examples{
var.names = LETTERS[1:20]
r("C", "F", var.names)
}
\author{
Dustin Fife
}
\seealso{
\code{\link{make.null}}, \code{\link{get.cols}}
}

