% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{report.fifer.anova}
\alias{report}
\alias{report.default}
\alias{report.fifer.anova}
\title{Output APA style statistical significance from a fifer.anova object}
\usage{
\method{report}{fifer.anova}(object)
}
\arguments{
\item{object}{a fifer.anova object}
}
\description{
Output APA style statistical significance from a fifer.anova object
}

