% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.pred.R
\name{rfSensitivity}
\alias{rfSensitivity}
\title{Output accuracy, sensitivity, specificity, NPV, and PPV.}
\usage{
rfSensitivity(object)
}
\arguments{
\item{object}{An rfPred object}
}
\value{
A list containing the accuracy, sensitivity, etc.
}
\description{
Output accuracy, sensitivity, specificity, NPV, and PPV.
}

