% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxtemp.R
\name{roxtemp}
\alias{roxtemp}
\title{Generate a Roxygen Template}
\usage{
roxtemp()
}
\value{
\\item{}{}
}
\description{
Generate a Roxygen Template
}
\references{

}
\seealso{
\\code{\\link{MASS}}
}
\author{
Dustin Fife

Write the description here

Write the details here


}
