% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.lm}
\alias{visualize.lm}
\alias{visualize}
\alias{visualize.default}
\title{Visualize a fitted model}
\usage{
\method{visualize}{lm}(object, plot = "bivariate", linetype = "loess")
}
\arguments{
\item{object}{a object}
}
\description{
Visualize a fitted model
}
