% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\alias{visualize.lm}
\alias{visualize.default}
\title{Visualize a fitted model}
\usage{
visualize(object, ...)
}
\arguments{
\item{object}{a object}
}
\description{
Visualize a fitted model
}
